Add SubmissionGroup(Category) and refactor a bit
