+++ hub/src/main/kotlin/org/sourcegrade/lab/hub/domain/Assignment.kt
@@ -1,6 +1,5 @@
 package org.sourcegrade.lab.hub.domain
 
-import org.sourcegrade.lab.hub.db.DomainEntity
 import java.time.ZonedDateTime
 import java.util.UUID
 
@@ -10,4 +9,5 @@ data class Assignment(
     val description: String,
     val course: Course,
     val submissionDeadline: ZonedDateTime,
+    val submissionGroupCategory: SubmissionGroupCategory,
 ) : DomainEntity
