+++ hub/src/main/kotlin/org/sourcegrade/lab/hub/domain/DomainEntity.kt
@@ -1,4 +1,4 @@
-package org.sourcegrade.lab.hub.db
+package org.sourcegrade.lab.hub.domain
 
 import java.util.UUID
