+++ hub/src/main/kotlin/org/sourcegrade/lab/hub/domain/GradingRun.kt
@@ -1,6 +1,5 @@
 package org.sourcegrade.lab.hub.domain
 
-import org.sourcegrade.lab.hub.db.DomainEntity
 import java.util.UUID
 
 data class GradingRun(
