+++ hub/src/main/kotlin/org/sourcegrade/lab/hub/domain/Submission.kt
@@ -1,6 +1,5 @@
 package org.sourcegrade.lab.hub.domain
 
-import org.sourcegrade.lab.hub.db.DomainEntity
 import java.time.ZonedDateTime
 import java.util.UUID
 
@@ -8,6 +7,7 @@ data class Submission(
     override val id: UUID,
     val assignment: Assignment,
     val submitter: User,
+    val group: SubmissionGroup,
     val uploaded: ZonedDateTime,
     val gradingRuns: List<GradingRun>,
     val lastGradingRun: GradingRun,
