+++ hub/src/main/kotlin/org/sourcegrade/lab/hub/domain/SubmissionGroup.kt
@@ -2,8 +2,8 @@ package org.sourcegrade.lab.hub.domain
 
 import java.util.UUID
 
-data class Role(
+data class SubmissionGroup(
     override val id: UUID,
-    val scope: String,
-    val permissions: List<String>,
+    val users: List<User>,
+    val category: SubmissionGroupCategory,
 ) : DomainEntity
